shmem-conduit fixes
   * Add shmem-conduit to SUBDIRS, add ranlib fixes, suggest mpirun as default
     on cray X1 and MIN/MAX fixes to basic.h.
   * Unicos/mp (*cough* IRIX) now supports mmap, although we probably won't be
     using it for the GASNet segment as remote access is only enabled by the
     shmalloc() mallocator -- this is still pending right now as I havn't found
     a good way to estimate the max amount of allocatable shmalloc() bytes.
   * Small fixes to accomodate Cray compiler.

   * Add a new cray-specific test in contrib, which shows the overhead of the
     shmem layer for remote and local memory loads/stores.
      * remote store: shmem=3.77us,   assembly=1.9us
      * remote load:  shmem=12.074us, assembly=9.01us
      * local  store: shmem=3.67us,   assembly=1.908us
      * local  load:  shmem=4.87us,   assembly=2.461us

     These loads are blocking and the stores are not.    On average, we
     recover 2-3 microseconds by not issuing the shmem library call.
     If we manage to deciper how remote memory segments are mapped into
     each segment, we may be able to bypass shmem altogether for puts/gets.

+++ shmem-conduit/gasnet_core.c
@@ -1,6 +1,6 @@
 /*  $Archive:: /Ti/GASNet/shmem-conduit/gasnet_core.c                  $
- *     $Date: 2003/11/12 08:56:04 $
- * $Revision: 1.1.2.2 $
+ *     $Date: 2003/11/13 08:06:11 $
+ * $Revision: 1.1.2.3 $
  * Description: GASNet shmem conduit Implementation
  * Copyright 2002, Dan Bonachea <bonachea@cs.berkeley.edu>
  * Terms of use are as specified in license.txt
@@ -545,7 +545,7 @@ gasnetc_AMPoll() {
 	    GASNETC_AMHEADER_UNPACK(
 		gasnetc_amq_reqs[idx].header,
 		amhdr.reqrep, amhdr.type, amhdr.numargs, 
-		amhdr.handler, (uint32_t) amhdr.pe);
+		amhdr.handler, amhdr.pe);
 
 #if 0
 	    printf("%d> AMPoll: received handler id %d from %d in slot %d\n",
